Use `flutter gen-l10n command` instead of `flutter intl plugin` to generate localizations
