+++ example/lib/screens/settings.dart
@@ -1,6 +1,5 @@
 // ignore_for_file: constant_identifier_names
 
-import 'package:fluent_ui/generated/l10n.dart';
 import 'package:flutter/foundation.dart';
 
 import 'package:fluent_ui/fluent_ui.dart';
@@ -91,7 +90,7 @@ class Settings extends ScrollablePage {
     const spacer = SizedBox(height: 10.0);
     const biggerSpacer = SizedBox(height: 40.0);
 
-    final supportedLocales = const AppLocalizationDelegate().supportedLocales;
+    const supportedLocales = FluentLocalizations.supportedLocales;
     final currentLocale =
         appTheme.locale ?? Localizations.maybeLocaleOf(context);
