+++ l10n.yaml
@@ -1,0 +1,10 @@
+arb-dir: lib/l10n
+output-class: FluentLocalizations
+output-localization-file: fluent_localizations.dart
+output-dir: lib/l10n/generated
+template-arb-file: intl_en.arb
+preferred-supported-locales: ["en"]
+synthetic-package: false
+nullable-getter: false
+format: true
+untranslated-messages-file: lib/l10n/untranslated.json
