--- lib/l10n/README.md
@@ -1,1 +1,0 @@
-Every `.arb` localization must have a corresponding entry in `localization.dart:defaultSupportedLocales` list.
