+++ lib/l10n/intl_en.arb
@@ -1,34 +1,127 @@
 {
   "@@locale": "en",
   "backButtonTooltip": "Back",
+  "@backButtonTooltip": {
+    "description": "The tooltip for the back button on [NavigationAppBar]."
+  },
   "closeButtonLabel": "Close",
+  "@closeButtonLabel": {
+    "description": "Label for \"close\" buttons and menu items."
+  },
   "searchLabel": "Search",
+  "@searchLabel": {
+    "description": "Label for \"search\" text fields."
+  },
   "closeNavigationTooltip": "Close Navigation",
+  "@closeNavigationTooltip": {
+    "description": "The tooltip for the toggle navigation button."
+  },
   "openNavigationTooltip": "Open Navigation",
+  "@openNavigationTooltip": {
+    "description": "The tooltip for the toogle navigation button."
+  },
   "clickToSearch": "Click to search",
+  "@clickToSearch": {
+    "description": "The tooltip for the \"Click to Search\" button."
+  },
   "modalBarrierDismissLabel": "Dismiss",
+  "@modalBarrierDismissLabel": {
+    "description": "Label read out by accessibility tools (TalkBack or VoiceOver) for a modal barrier to indicate that a tap dismisses the barrier. A modal barrier can for example be found behind an alert or popup to block user interaction with elements behind it."
+  },
   "minimizeWindowTooltip": "Minimize",
+  "@minimizeWindowTooltip": {
+    "description": "The tooltip used by the \"Minimize\" button on desktop windows."
+  },
   "restoreWindowTooltip": "Restore",
+  "@restoreWindowTooltip": {
+    "description": "The tooltip used by the \"Restore\" button on desktop windows."
+  },
   "closeWindowTooltip": "Close",
+  "@closeWindowTooltip": {
+    "description": "The tooltip used by the \"Close\" button on desktop windows."
+  },
   "dialogLabel": "Dialog",
+  "@dialogLabel": {
+    "description": "The dialog label."
+  },
   "cutActionLabel": "Cut",
+  "@cutActionLabel": {
+    "description": "The label for the cut action on the text selection controls."
+  },
   "copyActionLabel": "Copy",
+  "@copyActionLabel": {
+    "description": "The label for the copy action on the text selection controls."
+  },
   "pasteActionLabel": "Paste",
+  "@pasteActionLabel": {
+    "description": "The label for the paste button on the text selection controls."
+  },
   "selectAllActionLabel": "Select all",
+  "@selectAllActionLabel": {
+    "description": "The label for the select all button on the text selection controls."
+  },
   "newTabLabel": "Add new tab",
+  "@newTabLabel": {
+    "description": "The label used by [TabView]'s new button."
+  },
   "closeTabLabelSuffix": "Close tab",
+  "@closeTabLabelSuffix": {
+    "description": "The label used by [TabView]'s close button."
+  },
   "scrollTabBackwardLabel": "Scroll tab list backward",
+  "@scrollTabBackwardLabel": {
+    "description": "The label used by [TabView]'s scroll backward button."
+  },
   "scrollTabForwardLabel": "Scroll tab list forward",
+  "@scrollTabForwardLabel": {
+    "description": "The label used by [TabView]'s scroll forward button."
+  },
   "noResultsFoundLabel": "No results found",
+  "@noResultsFoundLabel": {
+    "description": "The label used by [AutoSuggestBox] when the results can't be found."
+  },
   "copyActionTooltip": "Copy the selected content to the clipboard",
+  "@copyActionTooltip": {
+    "description": "The tooltip for the copy action on the text selection controls."
+  },
   "cutActionTooltip": "Remove the selected content and put it in the clipboard",
+  "@cutActionTooltip": {
+    "description": "The tooltip for the cut action on the text selection controls."
+  },
   "pasteActionTooltip": "Inserts the contents of the clipboard at the current location",
+  "@pasteActionTooltip": {
+    "description": "The tooltip for the paste action on the text selection controls."
+  },
   "selectAllActionTooltip": "Select all content",
+  "@selectAllActionTooltip": {
+    "description": "The tooltip for the select all action on the text selection controls."
+  },
   "hour": "hour",
+  "@hour": {
+    "description": "The text used by [TimePicker] for the hour field."
+  },
   "minute": "minute",
+  "@minute": {
+    "description": "The text used by [TimePicker] for the minute field."
+  },
   "am": "AM",
+  "@am": {
+    "description": "The text used by [TimePicker] for the AM field."
+  },
   "pm": "PM",
+  "@pm": {
+    "description": "The text used by [TimePicker] for the PM field."
+  },
   "month": "month",
+  "@month": {
+    "description": "The text used by [DatePicker] for the month field"
+  },
   "day": "day",
-  "year": "year"
+  "@day": {
+    "description": "The text used by [DatePicker] for the day field"
+  },
+  "year": "year",
+  "@year": {
+    "description": "The text used by [DatePicker] for the year field"
+  }
 }
